Add sample of list of charted/uncharted songs
This commit adds a very small set of 5 charted and 5 uncharted songs
found manually to some files for testing data extraction. The songs are
listed in the format artist,track. These are just for testing extracting
the features for a list of songs to a csv for further processing and
will be updated to include more songs in the future.