Update Docker configuration
This commit removes the creation of a non-root user from the Dockerfile
due to problems with accessing particular directories. This may be
reversed in the future but for simplicity it has been removed. Also
the directory has been renamed from /code to /app to distinguish between the
app and the code subdirectory. The data subsdirectory has also been
mounted from the host machine.