Extract features and classification into csv
This commit uses a list of songs provided in the format artist, track in
a csv file(s) to extract a number of features from the Spotify API and
get the classification as to whether the song has charted from the
officialcharts website through scraping and then places all of this info
in the data.csv file to enable analysis. The list of features is not the
final list of features but offers an insight into how they can can be
placed into a data file for ML analysis later.