Extract song features and classifications
This commit adds the resultant files of using the songs.csv file to
create two csv files consisting of the songs with their respective
features and classifications. The data has been split between two files,
charted and uncharted, to enable even amounts of each to be used in
training/testing.