Add logistic regression
This commit adds a logistic regression classifier with cross validation
to select the two hyperparameters, q and C. The plots from the cross
validation are stored in a new plots directory. The performance of the
logistic regression appears to be no better than the baseline currently
but there is not enough data to assess this properly.