Use temporary data workaround
Previously, songs in the top 40 were considered 'charted' while thode
from 40-100 were considered 'uncharted'. This was leading to poor
results as the differences are not clear enough. For example, a track at
40 and a track at 41 won't be hugely different. For now, all songs in
the top 100 are being considered charted and an equal number of tracks
are being taken from the original recs_no_duplicates file as the
uncharted cohort. This provides substantially more data but should be
changed to use tracks that are NOT FOUND (or outside the top 100) in web scraping the officialcharts.com