Feature selection
This commit analyses the features available and considers their
correlation with both the target classification and with other features
to enable appropriate features to be selected. At the moment, the
features energy, valence, instrumentalness and danceability are included
in the features. Many of the others are insigniifcant (e.g. duration) or
have very high correlations with features that are already included
(e.g. loudness is highly correlated with energy).